Better sid2midi function with note namepsace.

+++ SID.h
@@ -119,7 +119,7 @@ class SID
   void loadPiano(uint8_t voice);
   void configVoice(uint8_t voice, uint8_t waveform, uint16_t pulsewidth /*12bit*/, uint8_t attackDecay, uint8_t sustainRelease, uint8_t releasePoint);
 
-  uint16_t  midi2Sid(int midin);
+  uint16_t  midi2Sid(uint8_t midin);
   void play(uint8_t voice, uint16_t freq);
   
   // library-accessible "private" interface
